First commit of working prototype
Enemies spawn on the right in waves and move towards the base. Click
"Fence" or "Turret" to switch item to build, activate build with B.
Move the object with the mouse, and scrollwheel to rotate. Click the
mouse button to place the object. When enemies are killed, the kill
counter goes up. Enemies that reach the base are deleted. If the path
to the base is blocked, the enemy will repath to the nearest building
and attack until a path is cleared.

Code and scene still need some cleanup, perhaps some refactoring.
Assets are mostly default objects, no sound or menus yet.